# Conflicts:
#	ssafit/src/main/java/com/ssafy/ssafit/domain/SubUser.java

+++ ssafit/src/main/java/com/ssafy/ssafit/service/SubUserServiceImpl.java
@@ -1,16 +1,24 @@
 package com.ssafy.ssafit.service;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import javax.validation.ConstraintViolationException;
+
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.ssafy.ssafit.domain.ApiResMessage;
 import com.ssafy.ssafit.domain.GetCt;
 import com.ssafy.ssafit.domain.MainUser;
 import com.ssafy.ssafit.domain.SubUser;
 import com.ssafy.ssafit.repository.GetCtRepository;
+import com.ssafy.ssafit.repository.MainuserRepository;
 import com.ssafy.ssafit.repository.SubUserRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -21,23 +29,51 @@ public class SubUserServiceImpl implements SubUserService {
 
 	private final SubUserRepository subUserRepository;
 	private final GetCtRepository getCtRepository;
+	private final MainuserRepository mainuserRepository;
 	
 	// 서브 계정 추가
 	@Override
-	public void addSubUser(SubUser subUser) {
+	public void addSubUser(Map<String, String> subUser) {
 		try {
-			subUserRepository.save(subUser);
-		} catch (Exception e) {
+			subUserRepository.save((SubUser.builder()
+					.nickName(subUser.get("nickName"))
+					.age(Integer.parseInt(subUser.get("age")))
+					.weight(Integer.parseInt(subUser.get("weight")))
+					.tall(Integer.parseInt(subUser.get("tall")))
+					.mainUser(mainuserRepository.findById(Long.parseLong(subUser.get("id"))).get())
+					.build()
+					));
+		} catch(ConstraintViolationException e) {
+			throw e;
+		}
+		catch (DataIntegrityViolationException e) {
+			throw e;
+		}
+		catch (Exception e) {
 			throw e;
 		}
 	}
 	
 	// 서브 계정(자녀) 계정 목록 조회
 	@Override
-	public List<SubUser> getMySubUserList(MainUser mainuser) {
-		List<SubUser> result;
+	public List<Map<String, Object>> getMySubUserList(long id) {
+		List<Map<String, Object>> result = new ArrayList<Map<String,Object>>();
 		try {
-			result = subUserRepository.findByMainUser(mainuser).orElse(null);
+			MainUser m = mainuserRepository.findById(id).get();
+			List<SubUser> temp = subUserRepository.findByMainUser(m).orElse(null);
+			if(temp != null) {
+				for(SubUser su :  temp) {
+					Map<String, Object> obj = new HashMap<String, Object>();
+					obj.put("sid", su.getSid());
+					obj.put("nickName", su.getNickName());
+					obj.put("age", su.getAge());
+					obj.put("weight", su.getWeight());
+					obj.put("tall", su.getTall());
+					
+					result.add(obj);
+				}
+			}
+			
 		} catch (Exception e) {
 			throw e;
 		}	
@@ -58,7 +94,7 @@ public class SubUserServiceImpl implements SubUserService {
 				obj.put("tall", subUser.getTall());
 				obj.put("weight", subUser.getWeight());
 				obj.put("nickName", subUser.getNickName());
-				obj.put("mainuser", subUser.getMainUser());
+				obj.put("id", subUser.getMainUser().getId());
 			}	
 		} catch (Exception e) {
 			throw e;
@@ -68,14 +104,14 @@ public class SubUserServiceImpl implements SubUserService {
 
 	// 자녀 계정 정보 수정
 	@Override
-	public void modifySubUser(SubUser subUser) {
+	public void modifySubUser(Map<String, String> subUser) {
 		try {
-			Optional<SubUser> s = subUserRepository.findBySid(subUser.getSid());
-			s.ifPresent(updatedSub ->{
-				updatedSub.setNickName(subUser.getNickName());
-				updatedSub.setAge(subUser.getAge());
-				updatedSub.setTall(subUser.getTall());
-				updatedSub.setWeight(subUser.getWeight());
+			Optional<SubUser> s = subUserRepository.findBySid(Long.parseLong(subUser.get("sid")));
+			s.ifPresent(updatedSub -> {
+				updatedSub.setNickName(subUser.get("nickName"));
+				updatedSub.setAge(Integer.parseInt(subUser.get("age")));
+				updatedSub.setTall(Integer.parseInt(subUser.get("tall")));
+				updatedSub.setWeight(Integer.parseInt(subUser.get("weight")));
 				subUserRepository.save(updatedSub);
 			});
 		} catch (Exception e) {
@@ -83,13 +119,24 @@ public class SubUserServiceImpl implements SubUserService {
 		}
 	}
 	
+	// 자녀 계정 삭제
 	@Override
 	public void deleteSub(long sid) {
-		SubUser su = subUserRepository.findById(sid).orElse(null);
-		GetCt gc = su.getCid();
-		su.setCid(null);
-		getCtRepository.delete(gc);
-		subUserRepository.deleteById(su.getSid());
-
+		try {
+			SubUser su = subUserRepository.findBySid(sid).get();
+			if(su != null) {
+				GetCt gc = su.getCid();
+				if(gc != null) getCtRepository.delete(gc);
+				su.setCid(null);
+				subUserRepository.deleteById(sid);
+			}
+		}
+		catch (Exception e) {
+			throw e;
+		}
 	}
+	
+	// 캐릭터 변경
+		
+	// 획득한 캐릭터 목록 조회
 }
